+++ pyramid_signed_params/__init__.py
@@ -3,6 +3,8 @@
 """
 from __future__ import absolute_import
 
+import warnings
+
 from .interfaces import ISignedParamsService
 
 
@@ -14,6 +16,14 @@ def includeme(config):
     settings = config.get_settings()
     if 'pyramid_signed_params.secrets' in settings:
         config.include('pyramid_signed_params.jwt_signer')
+    try:
+        config.find_service_factory(ISignedParamsService)
+    except ValueError:
+        warnings.warn(
+            "No service has been configured ISignedParamsService. "
+            "The request methods added by pyramid_signed_params will not "
+            "work without one. (Consider setting "
+            "pyramid_signed_params.secrets in your .ini file.)")
 
 
 def signed_params(request):